requirements.api.txt
fastapi==0.111.0
uvicorn==0.30.0
pydantic==2.7.0
langgraph==0.2.28
langchain-ollama==0.2.3
langchain-core==0.3.33
requests==2.32.2
mlflow==2.13.0
python-dotenv>=1.0.0
pyarrow>=4.0.0,<16